Refactored collections and lookups
