+++ src/glod/db/account.py
@@ -5,8 +5,7 @@ __copyright__ = 'Copyright(c) Gordon Elliott 2017'
 
 from a_tuin.db import TableMap, PagedQuery, InstanceQuery
 
-from glod.model.account import Account, AccountStatus
-from glod.model.account_collection import AccountCollection
+from glod.model.account import Account, AccountStatus, AccountCollection
 
 from glod.db.db_column_type_map import DB_COLUMN_TYPE_MAP
