+++ src/glod/db/counterparty.py
@@ -5,8 +5,7 @@ __copyright__ = 'Copyright(c) Gordon Elliott 2017'
 
 from a_tuin.db import RelationMap, TableMap, PagedQuery, InstanceQuery
 
-from glod.model.counterparty import Counterparty, StandingOrderDonor
-from glod.model.counterparty_collection import CounterpartyCollection
+from glod.model.counterparty import Counterparty, StandingOrderDonor, CounterpartyCollection
 from glod.model.references import counterparty__parishioner
 
 from glod.db.db_column_type_map import DB_COLUMN_TYPE_MAP
