--- src/glod/model/counterparty_collection.py
@@ -1,11 +1,0 @@
-__copyright__ = 'Copyright(c) Gordon Elliott 2017'
-
-""" 
-"""
-
-
-from a_tuin.metadata.collection import Collection
-
-
-class CounterpartyCollection(Collection):
-    pass
