--- src/glod/model/nominal_account_collection.py
@@ -1,9 +1,0 @@
-__copyright__ = 'Copyright(c) Gordon Elliott 2017'
-
-""" 
-"""
-
-from a_tuin.metadata import Collection
-
-class NominalAccountCollection(Collection):
-    pass
