+++ src/glod/model/subject.py
@@ -3,7 +3,7 @@ __copyright__ = 'Copyright(c) Gordon Elliott 2017'
 """
 """
 
-from a_tuin.metadata import StringField, ObjectFieldGroupBase
+from a_tuin.metadata import StringField, ObjectFieldGroupBase, Collection
 
 
 class Subject(ObjectFieldGroupBase):
@@ -13,3 +13,7 @@ class Subject(ObjectFieldGroupBase):
         StringField('select_vestry_summary'),
         StringField('easter_vestry_summary'),
     )
+
+
+class SubjectCollection(Collection):
+    pass
