--- src/glod/model/subject_collection.py
@@ -1,10 +1,0 @@
-__copyright__ = 'Copyright(c) Gordon Elliott 2017'
-
-""" 
-"""
-
-from a_tuin.metadata import Collection
-
-
-class SubjectCollection(Collection):
-    pass
